requirements.txt
librosa==0.10.2.post1
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84 
albumentations==1.4.14
pystoi==0.4.1
python_speech_features==0.6
pesq==0.0.4
fast_bss_eval==0.1.4
torch==2.1.0
torchaudio==2.1.0
torchlibrosa==0.1.0
torchnet==0.0.4
torchvision==0.16.0
soundfile==0.12.1
speechmos-0.0.1.1
thop-0.1.1.post2209072238